requirements.txt
Django==1.6
argparse==1.2.1
dj-database-url==0.2.2
dj-static==0.0.5
django-toolbelt==0.0.1
django-uuidfield==0.4.0
gunicorn==18.0
opbeat==1.1.2
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
